Refactored player tool into robot tool, replaced player container with player manager that has lookup for player script

+++ Assets/Scripts/UI/ToolTip/ToolTipController.cs
@@ -35,7 +35,9 @@ namespace UI.ToolTip {
             newToolTip.setText(text);
             newToolTip.transform.position = position+offset;
         }
-        public void HideToolTip() {
+        public void HideToolTip()
+        {
+            if (transform.childCount == 0) return;
             GlobalHelper.deleteAllChildren(transform);
         }
     }
